functions/requirements.txt
azure-functions==1.21.3
azure-search-documents==11.6.0
azure-ai-documentintelligence==1.0.0
openai==1.55.3
msgraph-sdk==1.9.0
azure-identity==1.19.0
azure-data-tables==12.6.0
PyMuPDF==1.25.1
python-docx==1.1.2
tiktoken==0.8.0
tenacity==8.5.0
